exercise-files/06/requirements.txt
python-dotenv==1.1.0
colorama==0.4.6
rich==13.3.4
openai==0.27.8

